separar los métodos de trabajo con modelos para que cargar y guardar no se superpusiera con los entrenadores
