+++ src/__init__.py
@@ -1,0 +1,5 @@
+"""Machine learning pipeline for gas turbine fault detection."""
+
+from .pipeline import run_pipeline
+
+__all__ = ['run_pipeline']
