+++ src/models/__init__.py
@@ -1,0 +1,9 @@
+"""Method to manage the models store in this folder."""
+
+from .manage_models import (
+    save_model,
+    load_model
+)
+
+__all__ = ['save_model', 'load_model']
+
