+++ src/training/__init__.py
@@ -1,10 +1,8 @@
 """Model training module for machine learning pipeline."""
 
-from .train_model import (
-    train_logistic_regression,
-    save_model,
-    load_model
+from .logistic_regression import (
+    train_logistic_regression
 )
 
-__all__ = ['train_logistic_regression', 'save_model', 'load_model']
+__all__ = ['train_logistic_regression']
